+++ dev/auto_tester/container_r_var/full_test.sh
@@ -7,8 +7,7 @@ cp CoRC/R/swig_wrapper.R .
 cp CoRC/inst/extdata/test_names.cps .
 
 # install CoRC and further dependencies
-Rscript -e "$R_OPTS remotes::install_local('CoRC/', INSTALL_opts = '--install-tests');"
-Rscript -e "CoRC::getCopasi(path = 'COPASI.so');"
+Rscript -e "$R_OPTS remotes::install_local('CoRC/', upgrade = 'never', INSTALL_opts = '--install-tests', configure.args = '--copasi-lib-path=/work/COPASI.so');"
 
 # Debug info
 Rscript -e "sessioninfo::session_info('CoRC');"